pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medvision-toolkit"
version = "0.1.0"
description = "AI-powered medical imaging analysis toolkit for Project MedVision"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "transformers>=4.40.0",
    "torch",
    "accelerate",
    "Pillow",
    "requests",
    "matplotlib",
    "ipykernel",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests",
    "types-Pillow",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
